[Geometry] Move public headers in to interface dir
