+++ Geometry/CSCGeometry/interface/CSCWireTopology.h
@@ -11,7 +11,7 @@
 
 #include "Geometry/CSCGeometry/interface/WireTopology.h"
 #include "Geometry/CSCGeometry/interface/CSCWireGeometry.h"
-#include "Geometry/CSCGeometry/src/CSCWireGrouping.h"
+#include "Geometry/CSCGeometry/interface/CSCWireGrouping.h"
 #include "Geometry/CSCGeometry/interface/CSCWireGroupPackage.h"
 #include <utility>  // for std::pair
