+++ Geometry/CSCGeometry/src/CSCGangedWireGrouping.h
@@ -12,7 +12,7 @@
  *
  */
 
-#include "Geometry/CSCGeometry/src/CSCWireGrouping.h"
+#include "Geometry/CSCGeometry/interface/CSCWireGrouping.h"
 #include <vector>
 
 class CSCGangedWireGrouping : public CSCWireGrouping {
